Update water layer mask to remove error
Mask was being deleted and causing error part way through inversion. I moved mask to self.path['mask_elastic'], so it should not be deleted anymore. I also now make the mask and apply the mask in the same place (function apply_mask_elastic), which is more easy to follow than making the mask in an init code and applying the same mask in a different script.