Temporary fix to h5 file component missing
Previous commit (August 11th) deleted h5 dataset during each iteration if it was present. The problem was that the observed dataset components wouldn't be known. Now I just get the component from the synthetic dataset. Not a permanent fix, but it allows us to store and use the h5 dataset.